% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIO.R
\name{calcIO}
\alias{calcIO}
\title{Calc Input Output}
\usage{
calcIO(
  subtype = c("input", "output", "trade"),
  ieaVersion = "default",
  corrected = FALSE
)
}
\arguments{
\item{subtype}{Data subtype. See default argument for possible values.}

\item{ieaVersion}{Release version of IEA data, either 'default'
(vetted and used in REMIND) or 'latest'.}

\item{corrected}{boolean indicating whether corrections should be applied to the data
after mapping}
}
\value{
IEA data as MAgPIE object aggregated to country level
}
\description{
Computes IEA-based model data for different "subtypes" by use of raw IEA "Energy Balances" data
and a mapping that corresponds to the structure of "products" and "flows" of IEA.
}
\details{
Mapping structure example: IEA product ANTCOAL used for IEA flow TPATFUEL, contributes via REMIND technology
coaltr for generating sesofos from pecoal (REMIND names)
}
\author{
Anastasis Giannousakis, Falk Benke
}
