% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrenergy-package.R
\docType{package}
\name{mrenergy-package}
\alias{mrenergy}
\alias{mrenergy-package}
\title{mrenergy: Preprocessing functions for REMIND and other energy models using landuse data}
\description{
Preprocessing functions for REMIND and other energy models (buildings, transport, industry) using landuse data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrenergy}
  \item Report bugs at \url{https://github.com/pik-piam/mrenergy/issues}
}

}
\author{
\strong{Maintainer}: Falk Benke \email{benke@pik-potsdam.de}

Authors:
\itemize{
  \item Falk Benke \email{benke@pik-potsdam.de}
  \item Robin Hasse
  \item Pascal Weigmann
}

}
