% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrenergy-package.R
\docType{package}
\name{mrenergy-package}
\alias{mrenergy}
\alias{mrenergy-package}
\title{mrenergy: the package provides energy related data via the madrat framework}
\description{
Provides useful functions and a common structure to all the input data required to run the REMIND of model input data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrenergy}
  \item Report bugs at \url{https://github.com/pik-piam/mrenergy/issues}
}

}
\author{
\strong{Maintainer}: Falk Benke \email{benke@pik-potsdam.de}

Authors:
\itemize{
  \item Falk Benke \email{benke@pik-potsdam.de}
}

}
